Fixed watch list buttons not displaying correctly
Fixed issue where the add to watch list and remove from watch list buttons are synced across users even if said user doesnt have that movie in their watch list.
THIS METHOD OF CHECKING IS DISCUSTING, i am not proud of the loops but it works and I would rather not have to recode the database again just to improve resource efficiency.